// music_app/types/album.ts
export interface Album {
  idAlbum: string;
  idArtist: string;
  idLabel: string;
  strAlbum: string;
  strAlbumStripped: string;
  strArtist: string;
  strArtistStripped: string;
  intYearReleased: string;
  strStyle: string;
  strGenre: string;
  strLabel: string;
  strReleaseFormat: string;
  intSales: string;
  strAlbumThumb: string;
  strAlbumThumbHQ: string | null;
  strAlbumBack: string | null;
  strAlbumCDart: string;
  strAlbumSpine: string | null;
  strAlbum3DCase: string | null;
  strAlbum3DFlat: string | null;
  strAlbum3DFace: string | null;
  strAlbum3DThumb: string | null;
  strDescriptionEN: string;
  strDescriptionDE: string | null;
  strDescriptionFR: string | null;
  strDescriptionCN: string | null;
  strDescriptionIT: string | null;
  strDescriptionJP: string | null;
  strDescriptionRU: string | null;
  strDescriptionES: string | null;
  strDescriptionPT: string | null;
  strDescriptionSE: string | null;
  strDescriptionNL: string | null;
  strDescriptionHU: string | null;
  strDescriptionNO: string | null;
  strDescriptionIL: string | null;
  strDescriptionPL: string | null;
  intLoved: string | null;
  intScore: string;
  intScoreVotes: string;
  strReview: string;
  strMood: string;
  strTheme: string;
  strSpeed: string;
  strLocation: string | null;
  strMusicBrainzID: string;
  strMusicBrainzArtistID: string;
  strAllMusicID: string | null;
  strBBCReviewID: string | null;
  strRateYourMusicID: string | null;
  strDiscogsID: string;
  strWikidataID: string | null;
  strWikipediaID: string | null;
  strGeniusID: string;
  strLyricWikiID: string | null;
  strMusicMozID: string | null;
  strItunesID: string | null;
  strAmazonID: string | null;
  strLocked: string;
}